Allow click + drag of the bounding box to edit its position
This commit sets up using the positions we read in from the data file only as
the initial state for the keyframe bounding boxes.  Then, when we click and drag
the bounding box, we edit the most recent keyframe's bounding box to have a new
position based on the position of the mouse pointer relative to where it was
at mousedown.