Add keyframe addition and deletion
I've added two buttons for adding and deleting a keyframe.  The addition button
adds a keyframe that starts at the current video position.  It defaults to (0,
400, 0, 400), which is maybe not optimal (should it default to the previous
keyframe's position?).  The deletion button deletes the keyframe prior to the
current video position (i.e. the one that describes the box currently on
screen).

I also fixed a small visual issue where the scrubber could go slightly off the
edge of the bar, and added sorting of keyframes based on time, just in case it
becomes possible to drag them out of order.